+++ ICD.Common.Utils.Tests/Xml/XmlUtilsTest.cs
@@ -56,12 +56,12 @@ namespace ICD.Common.Utils.Tests.Xml
 			{
 				reader.ReadToNextElement();
 
-				IcdXmlAttribute[] attributes = reader.GetAttributes().ToArray();
+				KeyValuePair<string, string>[] attributes = reader.GetAttributes().ToArray();
 
 				Assert.AreEqual(2, attributes.Length);
-				Assert.AreEqual("attr1", attributes[0].Name);
+				Assert.AreEqual("attr1", attributes[0].Key);
 				Assert.AreEqual("1", attributes[0].Value);
-				Assert.AreEqual("attr2", attributes[1].Name);
+				Assert.AreEqual("attr2", attributes[1].Key);
 				Assert.AreEqual("2", attributes[1].Value);
 			}
 		}