+++ CHANGELOG.md
@@ -5,6 +5,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+ - XML TryGetAttribute methods
+
+### Changed
+ - Performance improvements when working with xml attributes
+
+### Removed
+ - Removed IcdXmlAttribute
 
 ## [7.1.0] - 2018-11-08
 ### Added